+++ BovineLabs.Core.Editor/Dependency/DOTSSearchWindow.cs
@@ -2,6 +2,7 @@
 //     Copyright (c) BovineLabs. All rights reserved.
 // </copyright>
 
+#if UNITY_2023_3_OR_NEWER
 namespace BovineLabs.Core.Editor.Dependency
 {
     using System.Collections.Generic;
@@ -121,3 +122,4 @@ namespace BovineLabs.Core.Editor.Dependency
         }
     }
 }
+#endif