+++ BovineLabs.Core/Collections/CollectionCreator.cs
@@ -62,6 +62,43 @@ namespace BovineLabs.Core.Collections
             AllocatorManager.Free(allocator, hashMap);
         }
 
+        public static UnsafeParallelHashMap<TKey, TValue>* CreateParallelHashMap<TKey, TValue>(int initialCapacity, AllocatorManager.AllocatorHandle allocator)
+            where TKey : unmanaged, IEquatable<TKey>
+            where TValue : unmanaged
+        {
+            var hashMap = AllocatorManager.Allocate<UnsafeParallelHashMap<TKey, TValue>>(allocator);
+            *hashMap = new UnsafeParallelHashMap<TKey, TValue>(initialCapacity, allocator);
+            return hashMap;
+        }
+
+        public static void Destroy<TKey, TValue>(UnsafeParallelHashMap<TKey, TValue>* hashMap)
+            where TKey : unmanaged, IEquatable<TKey>
+            where TValue : unmanaged
+        {
+            var allocator = hashMap->m_AllocatorLabel;
+            hashMap->Dispose();
+            AllocatorManager.Free(allocator, hashMap);
+        }
+
+        public static UnsafeParallelMultiHashMap<TKey, TValue>* CreateParallelMultiHashMap<TKey, TValue>(
+	        int initialCapacity, AllocatorManager.AllocatorHandle allocator)
+	        where TKey : unmanaged, IEquatable<TKey>
+	        where TValue : unmanaged
+        {
+	        var hashMap = AllocatorManager.Allocate<UnsafeParallelMultiHashMap<TKey, TValue>>(allocator);
+	        *hashMap = new UnsafeParallelMultiHashMap<TKey, TValue>(initialCapacity, allocator);
+	        return hashMap;
+        }
+
+        public static void Destroy<TKey, TValue>(UnsafeParallelMultiHashMap<TKey, TValue>* hashMap)
+	        where TKey : unmanaged, IEquatable<TKey>
+	        where TValue : unmanaged
+        {
+	        var allocator = hashMap->m_AllocatorLabel;
+	        hashMap->Dispose();
+	        AllocatorManager.Free(allocator, hashMap);
+        }
+
         public static UnsafeQueue<T>* CreateQueue<T>(AllocatorManager.AllocatorHandle allocator)
             where T : unmanaged
         {